// FE.Core.FW/src/apis/models/TargetMappingModel.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type TargetMappingModel = {
  id?: string;
  year?: number;
  targetId?: string | null;
  productTypeId?: string | null;
  productTypeName?: string | null;
  customerCategoryId?: string | null;
  customerCategoryName?: string | null;
  jan?: number;
  quantityJan?: number;
  feb?: number;
  quantityFeb?: number;
  mar?: number;
  quantityMar?: number;
  apr?: number;
  quantityApr?: number;
  may?: number;
  quantityMay?: number;
  jun?: number;
  quantityJun?: number;
  july?: number;
  quantityJuly?: number;
  aug?: number;
  quantityAug?: number;
  sep?: number;
  quantitySep?: number;
  oct?: number;
  quantityOct?: number;
  nov?: number;
  quantityNov?: number;
  dec?: number;
  quantityDec?: number;
  total?: number;
  totalQuantity?: number;
  createdByUserId?: string;
  lastModifiedByUserId?: string;
  lastModifiedOnDate?: string;
  createdOnDate?: string;
};